Refactor PermissionPane prior to Redesign.
Moved PermissonPane to own class out of SettingsController,
moved parts of view-creation to fxml
