+++ src/main/java/org/openbase/bco/bcozy/permissions/package-info.java
@@ -1,0 +1,7 @@
+/**
+ *
+ * This package contains anything related with permissions.
+ *
+ * @author vdasilva
+ */
+package org.openbase.bco.bcozy.permissions;
