+++ src/main/resources/view/permissions/UnitPermissionView.fxml
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.CheckBox?>
+<?import javafx.scene.control.ChoiceBox?>
+<?import javafx.scene.control.TableColumn?>
+<?import javafx.scene.control.TableView?>
+<?import javafx.scene.layout.AnchorPane?>
+
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.shape.Box?>
+<?import javafx.scene.layout.HBox?>
+<?import org.openbase.bco.bcozy.view.ObserverButton?>
+<AnchorPane xmlns="http://javafx.com/javafx/8.0.111" xmlns:fx="http://javafx.com/fxml/1"
+            fx:controller="org.openbase.bco.bcozy.permissions.UnitPermissionController"
+            styleClass="detail-menu">
+
+    <TableView fx:id="permissionsTable" layoutX="4.0" layoutY="7.0" prefHeight="243.0" prefWidth="250.0"
+               AnchorPane.bottomAnchor="95.0" AnchorPane.leftAnchor="4.0" AnchorPane.rightAnchor="5.0"
+               AnchorPane.topAnchor="5.0">
+        <columns>
+            <TableColumn fx:id="usergroupColumn"/>
+            <TableColumn fx:id="permissionsColumn"/>
+        </columns>
+    </TableView>
+
+    <HBox fx:id="hbox" layoutX="8.0" layoutY="268.0" prefWidth="250.0">
+        <ChoiceBox fx:id="newGroupChoiceBox" layoutX="14.0" layoutY="262.0" prefHeight="27.0"/>
+        <CheckBox fx:id="readRights" mnemonicParsing="false"/>
+        <CheckBox fx:id="writeRights" mnemonicParsing="false"/>
+        <CheckBox fx:id="accessRights" mnemonicParsing="false"/>
+    </HBox>
+
+    <ObserverButton fx:id="saveRightsButton" identifier="save" layoutY="302.0" mnemonicParsing="false"
+                    onAction="#save"/>
+</AnchorPane>
